[#166761737] Order the list of organizations in alphabetical order
Moved selector to OldServicesHomeScreen.tsx
